permission middleware button hidden

+++ resources/views/company/list.blade.php
@@ -10,7 +10,9 @@
     <div class="alert alert-primary" role="alert">
         <h1>@lang('custom_title.company')</h1>
     </div>
-    <a class="my-2 btn btn-primary" href="/system-admin/company/new" role="button">+ @lang('custom_label.add_new')</a>
+    @if(in_array('company.register.form', $permission))
+        <a class="my-2 btn btn-primary" href="/system-admin/company/new" role="button">+ @lang('custom_label.add_new')</a>
+    @endif
     @if(session()->has('deleted'))
         <div class="alert alert-success">
             <strong>@lang('custom_message.deleted')</strong>
@@ -44,9 +46,15 @@
                         <td>{{$company->address}}</td>
                         <td>{{$company->business_plan_name}}</td>
                         <td>
-                            <a class="btn btn-info" href="/system-admin/company/detail/{{$company->id}}" role="button">@lang('custom_label.detail')</a>
-                            <a class="btn btn-primary" href="/system-admin/company/update/{{$company->id}}" role="button">@lang('custom_label.update')</a>
-                            <a onclick="return confirm('@lang('custom_message.confirm_delete')');" class="btn btn-danger" href="/system-admin/company/delete/{{$company->id}}" role="button">@lang('custom_label.delete')</a>
+                            @if(in_array('company.detail', $permission))
+                                <a class="btn btn-info" href="/system-admin/company/detail/{{$company->id}}" role="button">@lang('custom_label.detail')</a>
+                            @endif
+                            @if(in_array('company.update.form', $permission))
+                                <a class="btn btn-primary" href="/system-admin/company/update/{{$company->id}}" role="button">@lang('custom_label.update')</a>
+                            @endif
+                            @if(in_array('company.delete', $permission))
+                                <a onclick="return confirm('@lang('custom_message.confirm_delete')');" class="btn btn-danger" href="/system-admin/company/delete/{{$company->id}}" role="button">@lang('custom_label.delete')</a>
+                            @endif
                         </td>
                     </tr>
                 @endforeach
